fix: profile pic not synced after update
